+++ app/Http/Controllers/Privy/Dirjen/IndicatorController.php
@@ -17,6 +17,9 @@ class IndicatorController extends AdminController
     /**
      * Display a listing of the resource.
      *
+     * @param $planId
+     * @param $programId
+     * @param $targetId
      * @return \Illuminate\Http\Response
      */
     public function index($planId, $programId, $targetId)
@@ -40,17 +43,24 @@ class IndicatorController extends AdminController
      */
     public function create()
     {
-        //
+        if(\Gate::allows('read-only'))
+            abort(403);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
+     * @param $planId
+     * @param $programId
+     * @param $targetId
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request, $planId, $programId, $targetId)
     {
+        if(\Gate::allows('read-only'))
+            abort(403);
+
         $target = Target::find($targetId);
         $target->indicators()->save(new Indicator([
             'name'  => $request->get('name'),
@@ -76,11 +86,18 @@ class IndicatorController extends AdminController
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param $planId
+     * @param $programId
+     * @param $targetId
+     * @param $indicatorId
      * @return \Illuminate\Http\Response
+     * @internal param int $id
      */
     public function edit($planId, $programId, $targetId, $indicatorId)
     {
+        if(\Gate::allows('read-only'))
+            abort(403);
+
         return view('private.indicator.program_edit')
             ->with('id', [
                 'plan'  => $planId,
@@ -94,12 +111,19 @@ class IndicatorController extends AdminController
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param $planId
+     * @param $programId
+     * @param $targetId
+     * @param $indicatorId
      * @return \Illuminate\Http\Response
+     * @internal param int $id
      */
     public function update(Request $request, $planId, $programId, $targetId, $indicatorId)
     {
+        if(\Gate::allows('read-only'))
+            abort(403);
+
         $indicator = Indicator::find($indicatorId);
         $indicator->name = $request->get('name');
         $indicator->unit = $request->get('unit');
@@ -110,11 +134,18 @@ class IndicatorController extends AdminController
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param $planId
+     * @param $programId
+     * @param $targetId
+     * @param $indicatorId
      * @return \Illuminate\Http\Response
+     * @internal param int $id
      */
     public function destroy($planId, $programId, $targetId, $indicatorId)
     {
+        if(\Gate::allows('read-only'))
+            abort(403);
+
         return (int) Indicator::destroy($indicatorId);
     }
